// src/api/common/Authentication/types.ts
type BwType = InboundBwType | OutboundBwType

type Message<T extends BwType> = {
  bw: T
  id: number
}

type BrowserData = {
  browser: string
  browserVersion: string
  fp: string
  ip: string
  lang: string
  os: string
  osVersion: string
  time: number
  webIdentityKey: string
}

type InitMessage = Message<'init'> & { wp: BrowserData }

type LogoutMessage = Message<'logout'> & {
  cr: { apikey: string }
}

type ReconnectMessage = Message<'reconn'> & {
  wp: BrowserData
  cr: { apikey: string }
}

type TokenMessage = Message<'token'> & {
  username: string
  domain: string
  apikey: string
  lang: string
  mobileIdentityKey: string
  status: string
  enc: boolean
  userId: string
}

type ActivateMessage = Message<'activate'> & {
  cr: { apikey: string }
}

type StatisticsMessage = Message<'statistics'> & {
  message: {
    type: string
    osType: string
    msisdn: string
    appVersion: string
    webVersion: string
    time: number
  }
}

type PingMessage = Message<'ping'>
type TempMessage = Message<'temp'>